Move keys to the different tabs, update styles to better match new design
[#179940030]
[#180267346]
