--- css-modules/color-key.less
@@ -1,174 +1,0 @@
-@import "../css/variables.less";
-
-@topPos: @topBarHeight + 10px;
-@rightPos: 10px;
-
-.keyToggleButton {
-  position: absolute;
-  top: @topPos;
-  right: @rightPos;
-  background: white;
-  min-width: 72px;
-  display: flex;
-  border-radius: 5px;
-  box-shadow: 1px 1px 4px 0 rgba(0, 0, 0, 0.35);
-  border: solid 1px #797979;
-  .keyIconContainer {
-    background: #32a447;
-    width: 24px;
-    height: 24px;
-    border-radius: 50%;
-    :global(.material-icons) {
-      font-size: 20px;
-    }
-  }
-  &:hover {
-    background: #98d1a3 !important;
-  }
-  span {
-    margin-top: 2px;
-    color: white;
-  }
-  .label {
-    margin-left: 5px;
-    color: #434343;
-    font-size: 16px;
-  }
-}
-
-.colorKey {
-  color: #434343;
-  position: absolute;
-  top: 90px; // it used to be @topPos; interim change not to cover tabs. This whole component will be refactored.
-  right: @rightPos;
-  background: #fff;
-  min-width: 120px;
-  max-height: calc(100% - @bottomBarHeight - @topBarHeight - 20px);
-  overflow: auto;
-  padding: 0 5px 5px 5px;
-  text-align: right;
-
-  .closeIcon {
-    margin-top: 5px;
-    font-weight: bolder;
-    font-size: 20px;
-    color: #777777;
-    &:hover {
-      color: #bebebe;
-      cursor: pointer;
-    }
-  }
-
-  .keyTable {
-    color: #434343;
-    width: 100%;
-    margin-top: -15px;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-
-    th {
-      border-spacing: 10px 5px;
-      font-size: 16px;
-      font-weight: 400;
-      height: 25px;
-      padding: 5px;
-      padding-bottom: 0;
-      margin-top: 5px;
-      text-align: center;
-
-      &.subheader {
-        font-size: 1.2em;
-        height: 18px;
-      }
-    }
-
-    tr {
-      min-height: 23px;
-    }
-  }
-
-  .colorKeyContainer {
-    text-align: right;
-  }
-
-  .canvases {
-    display: inline-block;
-  }
-
-  .topo {
-    canvas {
-      height: 80px;
-      width: 22px;
-    }
-  }
-
-  .plate, .age {
-    canvas {
-      height: 80px;
-      width: 15px;
-    }
-  }
-
-  .labels {
-    display: inline-block;
-    vertical-align: top;
-  }
-
-  p {
-    margin-left: 5px;
-  }
-
-  .earthquakeKeyContainer {
-    .magnitudeText {
-      text-align: left;
-      padding-left: 2px;
-      padding-right: 18px;
-    }
-
-    .earthquakeDepth {
-      text-align: right;
-      padding-left: 3px;
-    }
-  }
-
-  .earthquakeMagnitudeGraphic {
-    width: 20px;
-  }
-
-  .volcanoKeyContainer {
-    .volcanoMarker {
-      width: 0;
-      height: 0;
-      border-style: solid;
-      border-width: 0 12px 20.8px 12px;
-      border-color: transparent transparent #ff6600 transparent;
-    }
-
-    .volcanoLabel {
-      line-height: 2em;
-      padding-left: 5px;
-    }
-  }
-
-  .rect {
-    width: 22px;
-    height: 20px;
-  }
-
-  .crossSectionKeyContainer {
-    text-align: right;
-  }
-
-  .crossSectionColor {
-    padding-left: 5px;
-  }
-}
-
-
-@media (max-width: 670px) {
-  .colorKey {
-    transform-origin: top right;
-    transform: scale(0.75, 0.75);
-  }
-}
