+++ css-modules/keys/map-type.less
@@ -1,0 +1,67 @@
+.mapType {
+  color: #434343;
+  background: #fff;
+  min-width: 120px;
+  overflow: auto;
+  text-align: right;
+  margin-bottom: 15px;
+  font-size: 14px;
+
+  .keyTable {
+    color: #434343;
+    width: 100%;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+
+    th {
+      font-size: 14px;
+      font-weight: bold;
+      margin: 5px 0 5px 0;
+      border-spacing: 10px 5px;
+      height: 25px;
+      padding-bottom: 0;
+      text-align: center;
+
+      &.subheader {
+        font-size: 1.2em;
+        height: 18px;
+      }
+    }
+
+    tr {
+      min-height: 23px;
+    }
+  }
+
+  .colorKeyContainer {
+    text-align: right;
+  }
+
+  .canvases {
+    display: inline-block;
+  }
+
+  .topo {
+    canvas {
+      height: 80px;
+      width: 22px;
+    }
+  }
+
+  .plate, .age {
+    canvas {
+      height: 80px;
+      width: 15px;
+    }
+  }
+
+  .labels {
+    display: inline-block;
+    vertical-align: top;
+  }
+
+  p {
+    margin-left: 5px;
+  }
+}
