+++ css-modules/keys/rock-types.less
@@ -1,4 +1,4 @@
-@import "../css/variables.less";
+@import "../../css/variables.less";
 
 .rockKey {
   color: #434343;
@@ -9,7 +9,7 @@
   .title {
     font-weight: bold;
     text-align: center;
-    margin-bottom: 5px;
+    margin: 5px 0 5px 0;
   }
 
   .container {
